requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
SQLAlchemy>=2.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
streamlit>=1.28.0
streamlit-autorefresh>=1.0.1
plotly>=5.17.0
rich>=13.6.0
tabulate>=0.9.0
pytz>=2023.3
httpx>=0.25.0
tenacity>=8.2.0
pytest>=7.4.0
pytest-mock>=3.12.0
joblib>=1.3.0
discord-webhook>=1.3.0
